address issues found in PR review
