+++ lib/hooks/Fetch.js
@@ -2,7 +2,7 @@
 
 import {createExcessiveUsageIdentifier} from '../excessiveUsageIdentification';
 import type {ObserveResourcePerformanceResult} from '../performanceObserver';
-import {addResourceTiming, addCorrelationHttpHeaders, captureHeaders} from './xhrHelpers';
+import {addResourceTiming, addCorrelationHttpHeaders} from './xhrHelpers';
 import {addCommonBeaconProperties} from '../commonBeaconProperties';
 import {observeResourcePerformance} from '../performanceObserver';
 import {now, generateUniqueId, matchesAny} from '../util';
@@ -85,7 +85,7 @@ export function instrumentFetch() {
       addCorrelationHttpHeaders(request.headers.append, request.headers, spanAndTraceId);
     }
 
-    captureHeaders(request.headers.forEach, request.headers, beacon);
+    captureHttpHeaders(request.headers, beacon);
 
     const performanceObserver = observeResourcePerformance({
       entryTypes: ['resource'],
@@ -98,7 +98,7 @@ export function instrumentFetch() {
     return originalFetch(request)
       .then(function(response) {
         beacon['st'] = response.status;
-        captureHeaders(response.headers.forEach, response.headers, beacon);
+        captureHttpHeaders(response.headers, beacon);
         // When accessing object properties as object['property'] instead of
         // object.property flow does not know the type and assumes string.
         // Arithmetic operations like addition are only allowed on numbers. OTOH,
@@ -168,4 +168,16 @@ function addGraphQlProperties(beacon: XhrBeacon, input, init) {
     }
   }
 
+}
+
+function captureHttpHeaders(headers, beacon) {
+  if (vars.headersToCapture.length === 0) {
+    return;
+  }
+
+  headers.forEach(function(value, name){
+    if(matchesAny(vars.headersToCapture, name)){
+      beacon['h_'+name.toLowerCase()] = value;
+    }
+  });
 }
