+++ lib/hooks/XMLHttpRequest.js
@@ -2,7 +2,7 @@
 
 import {createExcessiveUsageIdentifier} from '../excessiveUsageIdentification';
 import type {ObserveResourcePerformanceResult} from '../performanceObserver';
-import {addResourceTiming, addCorrelationHttpHeaders, captureXhrHeaders, convertHeader} from './xhrHelpers';
+import {addResourceTiming, addCorrelationHttpHeaders} from './xhrHelpers';
 import {addCommonBeaconProperties} from '../commonBeaconProperties';
 import {observeResourcePerformance} from '../performanceObserver';
 import {isAllowedOrigin} from '../allowedOrigins';
@@ -177,7 +177,7 @@ export function instrumentXMLHttpRequest() {
       beacon['d'] = Math.max(0, now() - (beacon['ts'] + vars.referenceTimestamp));
 
       if (vars.headersToCapture.length > 0) {
-        captureXhrHeaders(beacon, convertHeader(xhr.getAllResponseHeaders()));
+        captureHttpHeaders(beacon, xhr.getAllResponseHeaders());
       }
 
       if (state.performanceObserver && status > 0) {
@@ -274,4 +274,20 @@ export function instrumentXMLHttpRequest() {
     state.performanceObserver.onBeforeResourceRetrieval();
     return originalSend.apply(this, arguments);
   };
+}
+
+function captureHttpHeaders(beacon, headerString) {
+
+  const headers = {};
+  const lines = headerString.trim().split(/[\r\n]+/);
+  for( let i=0; i<lines.length; i++) {
+    const items = lines[i].split(': ', 2);
+    headers[items[0]] = items[1];
+  }
+
+  for(let key in headers) {
+    if(matchesAny(vars.headersToCapture, key)){
+      beacon['h_'+key.toLowerCase()] = headers[key];
+    }
+  }
 }
