+++ lib/hooks/xhrHelpers.js
@@ -1,6 +1,5 @@
 import {serializeEntryToArray} from '../resources/timingSerializer';
 import type {XhrBeacon} from '../types';
-import {matchesAny} from '../util';
 import vars from '../vars';
 
 /*
@@ -45,41 +44,4 @@ export function addCorrelationHttpHeaders(fn, ctx, traceId) {
   fn.call(ctx, 'X-INSTANA-T', traceId);
   fn.call(ctx, 'X-INSTANA-S', traceId);
   fn.call(ctx, 'X-INSTANA-L', '1,correlationType=web;correlationId=' + traceId);
-}
-
-export function captureHeaders(fn, ctx, beacon) {
-  if (vars.headersToCapture.length === 0) {
-    return;
-  }
-
-  fn.call(ctx, function(value, name){
-    if(matchesAny(vars.headersToCapture, name)){
-      beacon['h_'+name.toLowerCase()] = value;
-    }
-  });
-}
-
-export function captureXhrHeaders(beacon, headers) {
-  if (vars.headersToCapture.length === 0) {
-    return;
-  }
-
-  for(let key in headers) {
-    if(matchesAny(vars.headersToCapture, key)){
-      //may need check and validation on limits of header.
-      beacon['h_'+key.toLowerCase()] = headers[key];
-    }
-  }
-}
-
-export function convertHeader(headerString) {
-  const headers = {};
-  const lines = headerString.trim().split(/[\r\n]+/);
-  lines.forEach(function(line){
-    const items = line.split(': ');
-    const key = items[0];
-    const value = items.slice(1).join(': ');
-    headers[key] = value;
-  });
-  return headers;
 }
